Convert repo into engine module (#43)
* We don't need godot cpp here.

* Rename  to motion_matching_module

* Save wip.

* Do some cleanup.

* Apply pre-commit.

* Hide unused code.

* Try to fix warnings.

* Get physics process notification to the MMCharacter

* Don't build the editor for template_release and template_debug.

* Avoid variable shadowing.

* Style changes.

* Build for template release.

* Avoid typos.

* Generate doc templates.

* Avoid the build errors on different integer types.

* Null check skeleton before use.

* Build template targets.

* Avoid integer vs unsigned integer error.

* Avoid complaints about integer types.

* Size of an array size of int64.

* Fix error spam.

* Time is a double.

* Avoid shadowing variable names.

* Time is a double.

* Add a demo repository.

* Don't change the physics system.

---------

+++ src/features/mm_feature.cpp
@@ -1,6 +1,34 @@
-#include "mm_feature.h"
+/**************************************************************************/
+/*  mm_feature.cpp                                                        */
+/**************************************************************************/
+/*                         This file is part of:                          */
+/*                             GODOT ENGINE                               */
+/*                        https://godotengine.org                         */
+/**************************************************************************/
+/* Copyright (c) 2014-present Godot Engine contributors (see AUTHORS.md). */
+/* Copyright (c) 2007-2014 Juan Linietsky, Ariel Manzur.                  */
+/*                                                                        */
+/* Permission is hereby granted, free of charge, to any person obtaining  */
+/* a copy of this software and associated documentation files (the        */
+/* "Software"), to deal in the Software without restriction, including    */
+/* without limitation the rights to use, copy, modify, merge, publish,    */
+/* distribute, sublicense, and/or sell copies of the Software, and to     */
+/* permit persons to whom the Software is furnished to do so, subject to  */
+/* the following conditions:                                              */
+/*                                                                        */
+/* The above copyright notice and this permission notice shall be         */
+/* included in all copies or substantial portions of the Software.        */
+/*                                                                        */
+/* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,        */
+/* EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF     */
+/* MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. */
+/* IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY   */
+/* CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,   */
+/* TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE      */
+/* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                 */
+/**************************************************************************/
 
-#include <godot_cpp/core/error_macros.hpp>
+#include "mm_feature.h"
 
 MMFeature::MMFeature() {
 }
@@ -39,7 +67,7 @@ void MMFeature::denormalize(float* p_data) const {
 
 float MMFeature::compute_cost(const float* p_motion_data, const float* p_query_data) const {
     float cost = 0.0f;
-    for (int i = 0; i < get_dimension_count(); ++i) {
+    for (int64_t i = 0; i < get_dimension_count(); ++i) {
         float diff = p_motion_data[i] - p_query_data[i];
         cost += diff * diff;
     }
@@ -47,19 +75,19 @@ float MMFeature::compute_cost(const float* p_motion_data, const float* p_query_d
 }
 
 void MMFeature::_normalize_minmax(float* p_data) const {
-    for (int i = 0; i < get_dimension_count(); ++i) {
+    for (int64_t i = 0; i < get_dimension_count(); ++i) {
         p_data[i] = (p_data[i] - mins[i]) / (maxes[i] - mins[i]);
     }
 }
 
 void MMFeature::_denormalize_minmax(float* p_data) const {
-    for (int i = 0; i < get_dimension_count(); ++i) {
+    for (int64_t i = 0; i < get_dimension_count(); ++i) {
         p_data[i] = (p_data[i] * (maxes[i] - mins[i])) + mins[i];
     }
 }
 
 void MMFeature::_normalize_standard(float* p_data) const {
-    for (int i = 0; i < get_dimension_count(); ++i) {
+    for (int64_t i = 0; i < get_dimension_count(); ++i) {
         if (std_devs[i] < SMALL_NUMBER) {
             continue;
         }
@@ -68,7 +96,7 @@ void MMFeature::_normalize_standard(float* p_data) const {
 }
 
 void MMFeature::_denormalize_standard(float* p_data) const {
-    for (int i = 0; i < get_dimension_count(); ++i) {
+    for (int64_t i = 0; i < get_dimension_count(); ++i) {
         if (std_devs[i] < SMALL_NUMBER) {
             continue;
         }
